% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\name{DEGList}
\alias{DEGList}
\title{Identify differentially expressed gene list}
\usage{
DEGList(data, n1, n2, p.threshold = 0.05, padjust = "BH", mi = 5, NT = 100)
}
\arguments{
\item{data}{RNA-seq count data matrix, whose row contain genes and column contain samples}

\item{n1}{Control group}

\item{n2}{Case group}

\item{p.threshold}{adjusted pvalue by default=0.05}

\item{padjust}{select the p-value adjustment method by default="BH". choose the possible method "none", "BH", "fdr", "BY" and "holm". See \code{\link[stats]{p.adjust}} for more details.}

\item{mi}{maximum number of iterations to impute nullify values by default= 5}

\item{NT}{number of trees in each set to impute nullify values by default=100}
}
\value{
1. Gene.list: Differentially expressed genes list and 2. result: others output
}
\description{
DEGList is used to identify differentially expressed genes from RNA-seq count data.
}
\examples{
Brobustedger::DEGList(sampleDataOut,3,3,0.05) #Input RNA-seq count data
Brobustedger::DEGList(sampleDataMiss,3,3,0.05) #Input RNA-seq count data with missing or null value
}
\references{
Robinson MD, McCarthy DJ and Smyth GK (2010). edgeR: a Bioconductor package for differential expression analysis of digital gene expression data. Bioinformatics 26, 139-140

Stekhoven, D.J. and Buehlmann, P. (2012), 'MissForest - nonparametric missing value imputation for mixed-type data', Bioinformatics, 28(1) 2012, 112-118, doi: 10.1093/bioinformatics/btr597

George NI, Bowyer JF, Crabtree NM, Chang C-W (2015) An Iterative Leave-One-Out Approach to Outlier Detection in RNA-Seq Data. PLoS ONE 10(6): e0125224. https://doi.org/10.1371/journal.pone.0125224
}
\seealso{
Outliers identification function: \code{\link[Brobustedger]{iLOO}}, imputaion of missing/nullify function: \code{\link[Brobustedger]{NI}}, See details \code{\link[edgeR]{DGEList}}, \code{\link[missRanger]{missRanger}}, \code{\link[Brobustedger]{sampleDataMiss}} and \code{\link[Brobustedger]{sampleDataOut}}
}
\author{
Bandhan Sarker, Md. Matiur Rahaman, Muhammad Habibulla Alamin, Md. Ariful Islam, and Md. Nurul Haque Mollah
}
