% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\name{iLOO}
\alias{iLOO}
\title{Identify Outlier values from specific genes}
\usage{
iLOO(data)
}
\arguments{
\item{data}{RNA-seq count data matrix, whose row contain genes and column contain samples}
}
\value{
Output of the function identify gene specific outliers. Non-outlier values are displayed as "NA"
}
\description{
iLOO is used to identify gene specific outliers from RNA-seq count data.
}
\examples{
Brobustedger::iLOO(sampleDataOut) # Input RNA-seq count data
}
\seealso{
Differentially expressed gene list: \code{\link[Brobustedger]{DEGList}}, imputaion of missing/nullify function: \code{\link[Brobustedger]{NI}}, See details \code{\link[Brobustedger]{sampleDataMiss}} and \code{\link[Brobustedger]{sampleDataOut}}
}
