% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\name{NI}
\alias{NI}
\title{NULL value Imputation}
\usage{
NI(data, mi = 5, NT = 100)
}
\arguments{
\item{data}{RNA-seq count data matrix contains missing/Null value, whose row contain genes and column contain samples}

\item{mi}{maximum number of iterations to impute nullify values}

\item{NT}{number of trees in each set to impute nullify values}
}
\value{
imputation of missing/nullify data
}
\description{
NI is used to impute null/missing values in the RNA-seq count data.
}
\examples{
Brobustedger::NI(sampleDataMiss)  #Input RNA-seq count data with missing or null value
}
\seealso{
Differentially expressed gene list: \code{\link[Brobustedger]{DEGList}}, Outliers identification function: \code{\link[Brobustedger]{iLOO}}, See details \code{\link[missRanger]{missRanger}} , \code{\link[Brobustedger]{sampleDataMiss}} and \code{\link[Brobustedger]{sampleDataOut}}
}
