% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\docType{data}
\name{sampleDataMiss}
\alias{sampleDataMiss}
\title{Sample RNA-seq count data contains missing values}
\format{
RNAseq count \code{data_matrix},whose row contains 1000 genes and column contains 6 samples, where 3 samples are control groups and 3 sample are case groups
}
\usage{
sampleDataMiss
}
\description{
sample RNAseq count \code{data-matrix} contains missing values which is used for testing the package.
}
\examples{
data(sampleDataMiss)
}
\keyword{datasets}
