% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\docType{data}
\name{sampleDataOut}
\alias{sampleDataOut}
\title{Sample RNA-seq count data contains outlier values}
\format{
RNAseq count \code{data_matrix},whose row contains 1000 genes and column contains 6 samples, where 3 samples are control groups and 3 sample are case groups
}
\usage{
sampleDataOut
}
\description{
sample RNAseq count \code{data-matrix} contains outlier values which is used for testing the package.
}
\examples{
data(sampleDataOut)
}
\keyword{datasets}
