% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\name{DEGList}
\alias{DEGList}
\title{Identify differentially expressed gene list}
\usage{
DEGList(data, n1, n2, p.threshold = 0.05)
}
\arguments{
\item{data}{RNA-seq count data matrix, whose row contain genes and column contain samples}

\item{n1}{#Control group}

\item{n2}{#Case group}

\item{p.threshold}{adjusted pvalue by default=0.05}
}
\value{
Differentially expressed genes list
}
\description{
This function is used to identify differentially expressed genes from RNA-seq count data.
}
\examples{
Brobustedger::DEGList(sampleDataOut,3,3,0.05)
Brobustedger::DEGList(sampleDataMiss,3,3,0.05)
}
