% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\docType{data}
\name{sampleDataMiss}
\alias{sampleDataMiss}
\title{Sample RNA-seq count data with missing value}
\format{
RNAseq count \code{data_matrix},whose row contains 1000 genes and column contains 6 samples, where 3 control groups and 3 case groups
}
\usage{
sampleDataMiss
}
\description{
sample RNAseq count \code{data-matrix} with missing value that is used for testing
}
\examples{
data(sampleDataMiss)
}
\keyword{datasets}
