% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGList.R
\docType{data}
\name{sampleDataOut}
\alias{sampleDataOut}
\title{Sample RNA-seq count data with outliers value}
\format{
RNAseq count \code{data_matrix},whose row contains 1000 genes and column contains 6 samples, where 3 control groups and 3 case groups
}
\usage{
sampleDataOut
}
\description{
sample RNAseq count \code{data-matrix} with outlier’s value that is used for testing
}
\examples{
data(sampleDataOut)
}
\keyword{datasets}
